+++ src/main/java/ddd/caffeine/ratrip/module/place/application/PlaceService.java
@@ -18,10 +18,10 @@ import ddd.caffeine.ratrip.module.place.domain.Place;
 import ddd.caffeine.ratrip.module.place.domain.ThirdPartyDetailSearchOption;
 import ddd.caffeine.ratrip.module.place.domain.ThirdPartySearchOption;
 import ddd.caffeine.ratrip.module.place.domain.repository.PlaceRepository;
-import ddd.caffeine.ratrip.module.place.external.KakaoRegionApiClient;
-import ddd.caffeine.ratrip.module.place.external.dto.KakaoRegionResponse;
 import ddd.caffeine.ratrip.module.place.feign.PlaceFeignService;
+import ddd.caffeine.ratrip.module.place.feign.kakao.KakaoRegionApiClient;
 import ddd.caffeine.ratrip.module.place.feign.kakao.model.FeignPlaceModel;
+import ddd.caffeine.ratrip.module.place.feign.kakao.model.KakaoRegionResponse;
 import ddd.caffeine.ratrip.module.place.feign.naver.model.FeignBlogModel;
 import ddd.caffeine.ratrip.module.place.feign.naver.model.FeignImageModel;
 import ddd.caffeine.ratrip.module.place.presentation.dto.bookmark.BookmarkPlaceResponseDto;
