+++ src/main/java/ddd/caffeine/ratrip/module/place/feign/kakao/KakaoRegionApiClient.java
@@ -1,11 +1,11 @@
-package ddd.caffeine.ratrip.module.place.external;
+package ddd.caffeine.ratrip.module.place.feign.kakao;
 
 import org.springframework.cloud.openfeign.FeignClient;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestParam;
 
-import ddd.caffeine.ratrip.module.place.external.dto.KakaoRegionResponse;
+import ddd.caffeine.ratrip.module.place.feign.kakao.model.KakaoRegionResponse;
 
 @FeignClient(name = "kakaoRegionApi", url = "https://dapi.kakao.com")
 public interface KakaoRegionApiClient {
