+++ src/main/java/ddd/caffeine/ratrip/module/place/feign/kakao/model/KakaoRegionMeta.java
@@ -1,4 +1,4 @@
-package ddd.caffeine.ratrip.module.place.external.dto;
+package ddd.caffeine.ratrip.module.place.feign.kakao.model;
 
 public class KakaoRegionMeta {
 	private int total_count;
